move enums to their one place of usage
