+++ unreal_asset/src/enums.rs
@@ -45,44 +45,3 @@ pub enum ELifetimeCondition {
     CondNever = 15,
     CondMax = 16,
 }
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive)]
-#[repr(u8)]
-pub enum EBlueprintTextLiteralType {
-    /// Text is an empty string. The bytecode contains no strings, and you should use FText::GetEmpty() to initialize the FText instance.
-    Empty,
-    /// Text is localized. The bytecode will contain three strings - source, key, and namespace - and should be loaded via FInternationalization
-    LocalizedText,
-    /// Text is culture invariant. The bytecode will contain one string, and you should use FText::AsCultureInvariant to initialize the FText instance.
-    InvariantText,
-    /// Text is a literal FString. The bytecode will contain one string, and you should use FText::FromString to initialize the FText instance.
-    LiteralString,
-    /// Text is from a string table. The bytecode will contain an object pointer (not used) and two strings - the table ID, and key - and should be found via FText::FromStringTable
-    StringTableEntry,
-}
-
-#[derive(Debug, Hash, Copy, Clone, PartialEq, Eq, IntoPrimitive, TryFromPrimitive)]
-#[repr(i8)]
-pub enum TextHistoryType {
-    None = -1,
-    Base = 0,
-    NamedFormat,
-    OrderedFormat,
-    ArgumentFormat,
-    AsNumber,
-    AsPercent,
-    AsCurrency,
-    AsDate,
-    AsTime,
-    AsDateTime,
-    Transform,
-    StringTableEntry,
-    TextGenerator,
-    RawText, // Uncertain, Back 4 Blood specific serialization
-}
-
-impl Default for TextHistoryType {
-    fn default() -> Self {
-        Self::None
-    }
-}
